jisx0201.1976-0.enc: Add undefined mapping for 0x7f - 0xa0 range
The jisx0201 encoding has a hole in it in the 0x7f - 0xa0 range, see:
http://en.wikipedia.org/wiki/JIS_X_0201

The document for the format of the encoding files says:
"Codes not listed are assumed to map through the identity (i.e. to the same
 numerical value). In order to override this default mapping, you may specify
 a range of codes to be undefined by using an `UNDEFINE' line"

jisx0201.1976-0.enc was not doing this, causing all the codes on the range
to be mapped 1:1. This specifically is causing problems for ttmkfdir, which
checks that a font covers (most of) an encoding before adding a line for that
encoding to fonts.scale, and the hole not properly being undefined causes
this check to fail for atleast the Sazanami fonts.

Cc: Akira TAGOH <tagoh@redhat.com>
BugLink: https://bugzilla.redhat.com/show_bug.cgi?id=1009350
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>